requirements.txt
cassandra-driver
cassandra-migrate
flask